requirements.txt
fastapi
uvicorn
numpy
Pillow
onnxruntime
insightface
numpy
python-multipart